Add defect examples and move full framework project
