+++ integrationtest/NetFramework/ExampleLibrary/Class1.cs
