+++ integrationtest/NetFramework/ExampleLibrary/ExampleLibrary.csproj
