+++ integrationtest/NetFramework/ExampleLibrary/Properties/AssemblyInfo.cs
