+++ integrationtest/NetFramework/FullFrameworkApp.Test/FullFrameworkApp.Test.csproj
