+++ integrationtest/NetFramework/FullFrameworkApp.Test/Properties/AssemblyInfo.cs
