+++ integrationtest/NetFramework/FullFrameworkApp.Test/packages.config
