+++ integrationtest/NetFramework/FullFrameworkApp.Test/stryker-config.json
