+++ integrationtest/NetFramework/FullFrameworkApp/App.config
