+++ integrationtest/NetFramework/FullFrameworkApp/FullFrameworkApp.csproj
