+++ integrationtest/NetFramework/FullFrameworkApp/Person.cs
