+++ integrationtest/NetFramework/FullFrameworkApp/Properties/AssemblyInfo.cs
