+++ integrationtest/NetFramework/FullFrameworkApp/Translations.resx
