+++ integrationtest/NetFramework/FullFrameworkApp/packages.config
