--- integrationtest/TargetProjects/TargetProject/Defects/DummyCalc.cs
@@ -1,27 +1,0 @@
-using ExampleClassLibrary;
-
-namespace TargetProject.Defects
-{
-    public class DummyCalc
-    {
-        public int SomeCalc(int first, int second)
-        {
-            while (second > 0)
-            {
-                first = first + second + 1;
-                second--;
-            }
-            while (1 < 0)
-            {
-                // endless looping
-            }
-            return first;
-        }
-
-        public int Recursive(int n)
-        {
-            var recursive = new RecursiveMath();
-            return recursive.Fibonacci(n);
-        }
-    }
-}
