+++ integrationtest/TargetProjects/TargetProject/Defects/LamdaAssignment.cs
@@ -1,0 +1,11 @@
+// https://github.com/stryker-mutator/stryker-net/issues/339
+
+namespace TargetProject.Defects;
+
+public class LamdaAssignment
+{
+    public void DoIt()
+    {
+        Action act = () => Console.WriteLine("A", "B", "C");
+    }
+}
