+++ integrationtest/TargetProjects/TargetProject/Defects/LinqVariableName.cs
@@ -1,0 +1,13 @@
+// https://github.com/stryker-mutator/stryker-net/issues/250
+
+namespace TargetProject.Defects
+{
+    public class ShouldNotPlaceLinqMutant
+    {
+        public int SomeMethodWithOutParameter(out int Any)
+        {
+            Any = 1;
+            return 4;
+        }
+    }
+}
