--- integrationtest/TargetProjects/TargetProject/Defects/RecursiveCoalescingAssigment.cs
@@ -1,14 +1,0 @@
-using System.Collections.Generic;
-
-namespace TargetProject.Defects
-{
-    public class Myclass
-    {
-        public void SwitchExpression()
-        {
-            List<int> numbersA = null;
-            List<int> numbersB = null;
-            numbersA ??= numbersB ??= new List<int>();
-        }
-    }
-}
