+++ integrationtest/TargetProjects/TargetProject/Defects/WhileTrue.cs
@@ -1,0 +1,17 @@
+// https://github.com/stryker-mutator/stryker-net/issues/642
+
+namespace TargetProject.Defects;
+
+public class WhileTrue
+{
+    public static bool Loop(bool stop)
+    {
+        while (true)
+        {
+            if (stop)
+            {
+                return true;
+            }
+        }
+    }
+}
